requirements.txt
# 经济运行分析平台 - 依赖包列表
# 核心框架
streamlit>=1.50.0

# 数据处理
pandas>=2.3.0
numpy>=2.3.0
openpyxl>=3.1.0
xlrd>=2.0.0

# 科学计算与统计分析
scipy>=1.16.0
statsmodels>=0.14.0
scikit-learn>=1.7.0

# 可视化
matplotlib>=3.10.0
seaborn>=0.13.0
plotly>=6.3.0
altair>=5.5.0

# 系统与性能监控
psutil>=7.1.0

# 工具库
python-dateutil>=2.9.0
pytz>=2025.2
tqdm>=4.67.0
joblib>=1.5.0
bcrypt>=5.0.0

# 时间序列分析
dtaidistance>=2.3.0  # DTW动态时间规整（窗口约束有效）

# 其他依赖
colorama>=0.4.6
pillow>=11.3.0
